+++ aes-siv/CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- Bump `aes` crate dependency to v0.8 ([#387])
+- Bump `cipher` crate dependency to v0.4 ([#387])
+- Bump `cmac` crate dependency to v0.7 ([#387])
+- Bump `ctr` crate dependency to v0.9 ([#387])
+- Bump `pmac` crate dependency to v0.7 ([#387])
+- Bump `zeroize` crate dependency to v1.5 ([#387])
+
 ## 0.6.2 (2021-07-20)
 ### Changed
 - Pin `zeroize` dependency to v1.3 ([#349])