+++ chacha20poly1305/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- Bump `cipher` crate dependency to v0.4 ([#387])
+- Bump `chacha20` crate dependency to v0.9 ([#387])
+- Bump `zeroize` crate dependency to v1.5 ([#387])
+
 ## 0.9.0 (2021-08-29)
 ### Changed
 - Bump `chacha20` to v0.9: now a hard dependency ([#365])