+++ mgm/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- Bump `cipher` crate dependency to v0.4 ([#387])
+
 ## 0.4.6 (2021-09-01)
 ### Added
 - Target feature autodetection on x86(-64) targets ([#371])